requirements.txt
python-dotenv>=1.0.0
together>=0.2.8
Pillow>=10.0.0
PyQt6>=6.5.0
requests>=2.31.0
pystray>=0.19.4
